chore: add ci gate and clear lint and comment debt
Adds a pull-request workflow running fmt, clippy at -D warnings, and the
test suite, then clears the backlog it would fail on, including a
deny-by-default never_loop in the fusermount3 path.

+++ crates/onelf-rt/src/ephemeral.rs
@@ -22,7 +22,7 @@ fn tmpfs_size_for(pkg: &PackageData) -> u64 {
         .entries
         .iter()
         .flat_map(|e| e.blocks.iter())
-        .map(|b| u64::from(b.original_size))
+        .map(|b| b.original_size)
         .sum();
     let with_overhead = content_size + content_size / 4;
     with_overhead.max(8 * 1024 * 1024)
